Moved Options for BoolToConcept tool to tool hierarchy
