+++ plato.cabal
@@ -27,11 +27,11 @@ library
                       Tuura.Concept.FSM.Simulation,
                       Tuura.Concept.FSM.Translation,
                       Tuura.Boolean.Parser,
-                      Tuura.Boolean.Options,
                       Tuura.Boolean,
                       Tuura.Plato.Translate.Options,
                       Tuura.Plato.Translate.Translation,
-                      Tuura.Plato.BoolToConcept
+                      Tuura.Plato.BoolToConcept,
+                      Tuura.Plato.BoolToConcept.Options
     other-modules:    Tuura.Concept.Abstract,
                       Tuura.Concept.Circuit,
                       Tuura.Concept.Circuit.Basic,
